+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Released on Sunday, July 26 2026.
 - Added missing color declarations (e.g., `color-scheme`, `accent-color`)
 - Added missing compositing declarations (e.g., `backdrop-filter`, `isolation`)
 - Added missing shape declarations (e.g., `shape-outside`, `shape-rendering`)
+- Added missing auxiliary declarations (e.g., `counter-set`, `image-rendering`, `view-transition-name`)
 - Added more CSSOM possibilities and helpers (#6)
 - Added parts of recent color spec update such as `rgb` with spaces (#131)
 - Added now Color L4 parsing with `hsl`, `hwb`, `lab`, `lch`, `oklab`, and `oklch`